refactor(auto-run): decompose BatchRunnerModal into focused hooks
- Extract task-selection recommendation, Goal-Driven config, prompt
  composer state, Spec-Driven config, and worktree run-target
  selection into src/renderer/hooks/batch/.
- Each extraction is intended to preserve existing behavior; verified
  against the existing 131-test BatchRunnerModal suite and the
  broader 715-test batch hook suite.
- Fix a duplicate {{MAESTRO_CLI_PATH}} entry in
  shared/templateVariables.ts, found during manual testing and
  unrelated to the extraction.
