+++ src/renderer/hooks/batch/index.ts
@@ -111,6 +111,32 @@ export type {
 	UseWorktreeValidationDeps,
 } from './useWorktreeValidation';
 
+// Task/Document fresh-context recommendation engine
+export { useTaskSelectionRecommendation } from './useTaskSelectionRecommendation';
+export type {
+	UseTaskSelectionRecommendationReturn,
+	UseTaskSelectionRecommendationDeps,
+} from './useTaskSelectionRecommendation';
+
+// Goal-Driven Auto Run config (tab, goal, exit criteria, max iterations)
+export { useGoalDrivenConfig } from './useGoalDrivenConfig';
+export type { UseGoalDrivenConfigReturn, UseGoalDrivenConfigDeps } from './useGoalDrivenConfig';
+
+// Agent prompt state (text, saved/default flags, composer, template variables)
+export { usePromptComposerState } from './usePromptComposerState';
+export type {
+	UsePromptComposerStateReturn,
+	UsePromptComposerStateDeps,
+} from './usePromptComposerState';
+
+// Spec-Driven Auto Run config (documents, task counts, loop mode)
+export { useSpecDrivenConfig } from './useSpecDrivenConfig';
+export type { UseSpecDrivenConfigReturn, UseSpecDrivenConfigDeps } from './useSpecDrivenConfig';
+
+// Worktree run-target selection (which worktree a run dispatches to)
+export { useWorktreeRunTarget } from './useWorktreeRunTarget';
+export type { UseWorktreeRunTargetReturn, UseWorktreeRunTargetDeps } from './useWorktreeRunTarget';
+
 // Auto Run achievements/badges
 export { useAchievements, queueAchievement } from './useAchievements';
 export type {
